Set Deserialization on Unknown Properties as false
When the object mapper ties to deserialize the content in the response attributes, it is failing if there are unknown properties in the database which is not a desired behavior. Even if there are unknown properties it should ignore them. These unknown properties are because of the multiple changes in the DB added at the time of development. This will not occur with the release branch